feat: reverse second half of list

+++ 0x03-python-data_structures/lists.h
@@ -2,7 +2,6 @@
 #define LISTS_H
 
 #include <stdlib.h>
-#include <stdio.h>
 
 /**
  * struct listint_s - singly linked list
@@ -23,7 +22,7 @@ listint_t *add_nodeint_end(listint_t **head, const int n);
 void free_listint(listint_t *head);
 
 int is_palindrome(listint_t **head);
-size_t listint_len(listint_t *h);
+unsigned int listint_len(listint_t *h);
 listint_t *reverse_list(listint_t **h);
 
 #endif /* LISTS_H */
